+++ internal/cgen/drop-in/stb.c
@@ -116,9 +116,9 @@ wuffs_drop_in__stb__make_decoder(   //
         return wuffs_qoi__decoder__alloc_as__wuffs_base__image_decoder();
 #endif
 
-#if !defined(WUFFS_CONFIG__MODULES) || defined(WUFFS_CONFIG__MODULE__TGA)
+#if !defined(WUFFS_CONFIG__MODULES) || defined(WUFFS_CONFIG__MODULE__TARGA)
       case WUFFS_BASE__FOURCC__TGA:
-        return wuffs_tga__decoder__alloc_as__wuffs_base__image_decoder();
+        return wuffs_targa__decoder__alloc_as__wuffs_base__image_decoder();
 #endif
 
 #if !defined(WUFFS_CONFIG__MODULES) || defined(WUFFS_CONFIG__MODULE__THUMBHASH)